Minor clarifications to XEP-0198
Two things came up during community review of a patch to Openfire which are
not specified in the XEP. Both seem to be uncontentious, but will require
Council review, etc.

1) The specification makes no recommendation on what to do if the server
receives multiple <enable/> requests.

This is resolved in this instance by recommending (SHOULD, not MUST) that the
server generates a stream error, and stipulates that clients MUST NOT do this.

2) It was noted that redelivery attempts for stanzas require delay stamp
handling.

It is arguably not within this specification's remit to require this, however
it seems compelling that it be documented somewhere and this seems the best
place.