Change Carbons delivery rules to 'all IM traffic'
After discussion on the standards list at the time of Last Call, it
became clear that servers were not generally implementing Carbons as
specified. This patch updates the text to be more forgiving of such
discrepancies, and to try to reflect general consensus between
implementations of what should be eligible for carbons delivery.