XEP-0045: Business Rule for client-requested full resync
As discussed on the ML [0]: When a client re-joins a MUC where the MUC
service did not notice the client leaving, the client will end up with
an incomplete list of participants and no room subject.

This patch adds a sentence to the business rules section that explicitly
requires a MUC service to send a full sync whenever it receives a MUC
join presence (as opposed to the old groupchat protocol).

In my eyes, this is only a clarification of the existing standard
definition, so we do not need to bump the version.

Georg

[0] http://mail.jabber.org/pipermail/standards/2016-June/031180.html