XEP-0070: Mention a plaintext fallback
In the previous version, an entity requesting XEP-0070 validation could
add a <body/> element to the <message/> sent, which would be displayed
in any client (which may not support this XEP). If the client does not
support the XEP, and the entity does not support a plaintext fallback,
the client would be prompted to act, without any result being possible
(as the wording says that the <confirm/> element must be sent back as
well).

This change suggests that an entity adding a body should implement this
kind of fallback (which may be as dumb as "ok/nok", as long as the
<body/> element is clear enough).