XEP-0359: clarifications and removed 'overwrite' wording
This commit changes some of the wording to clarify that the by attribute
for messages tagged by the server should be either the account jid or
the room jid (as opposed to the server jid for example)

To further clarify this the examples (including the disco example) were
all changed to have the same jid/by-attribute. (A room in that example)

The business rules had some wording that an entity would overwrite the
ID in existing 'faked' stanza-ids instead of removing the faked and
re-adding a non-faked one. However that doesn't work for faked messages
that are not archived (don't have a stanza-id).

This commit also makes the XEP explicitly about message stanzas