XEP-0045: Added standardized roominfo fields for allowinvites and the room name.
The standard doesn’t commend muc services to set the identity name of a room to be the name configured in the room configuration. Some implementations set the identity name to the local part of the jid if the configured room name is empty. Therefore implemtations are unable to discover if a room name is empty (=not set) when running disco#info. To circumvent this and to let clients discover (reliably) if a name is set (or not!) this PR adds a standardized roominfo field for the name.

Discovering whether or not an occupant is allowed to invite other members is also useful (clients can hide the button if that’s not the case.)