XEP-0059: Add 'exact(-index)' attribute to RSM's <count/> and <first/>
RSM specifies that the number or position included in a RSM response
may be approximate. But some protocols benefit from knowledge whether
the value is exact or approximate. For example a MAM synchronization
algorithm, which syncs a slave archive, potentially having multiple
holes, from a master archive, may use an optimized algorithm if the
master archive returns exact counts.

This commit adds signaling to RSM that the values are exact and not
approximate in a backwards compatible fashion.