XEP-0198: Clarify the note about counters
The note was only 50% correct (the counter the note talks get
initialized and started when the entity receives <enable(d)/>, and not
when it is sent) and partly misleading (there are typically four
counters per stream). Most XEP-0198 Stream Management implementations
I am aware of didn't get the counters right at the first time. The
tricky part is to get the time right where the counters should be set
to zero and get started. The new note's text tries to help
implementers with this. As bonus, it is also shorter than the original
note.