XEP-0198: Clarify the note about counters
Many implementations have failed to get this right on their first
attempt, and so this text aims to clarify the logic required for
an implementation to correctly maintain its counters for a stream.

The text was 95% based on text originally proposed by Florian Schmaus,
but I find it a bit clearer and fixed some grammar in the process.