XEP-0030: Clarify 'disco#info' feature in 'disco#info' responses
Strictly speaking, every disco#info IQ result must contain the
disco#info feature. But looking at the existing examples in various
XEPs and implementations, this feature is usually omitted. Hence we
clarify in XEP-0030 that announcing the 'disco#info' feature is
optional, as support for it is implicitly acknowledged by returning a
result IQ to a 'disco#info' IQ request.