XEP-0001: add state transition Proposed -> Experimental
Rationale: Sometimes, XEPs linger in Last Call state. This is
probably for the same reason as XEPs which linger in Experimental
state (and then become Deferred). However, currently there is no
way to put Proposed (= Last Call) XEPs back into Experimental state
if one follows the XEP-0001 process by the letter.

This is unfortunate, because it puts council into the position to
either:

- Reject a XEP which is actually useful, but needs work before going
  to Draft, or
- Advance a XEP to Draft which isn’t ready yet, or
- Burn cycles of the Editors and Council on perpetuating a Last
  Call which won’t lead to a conclusion at the current time.

(The "burning" of "cycles" is mostly due to the noise created by
either officially extending the LC all the time, or having a zombie
Proposed XEP hanging around in the list and having to recognize it
as zombie and ignoring it.)