XEP-0308: Allow message correction in MUC.
Corrections done in MUC used to always be untrusted when coming from the
room’s history, because —according to this XEP— another entity could
have taken the same in-room full JID.

This addition makes the room responsible for rejecting such illegal
correction attempts, and advertise that in its disco#info, which clients
can then trust properly.

This also relaxes the rule on identity checking in non-anonymous rooms,
by making it rely on the real bare JID of the occupant.