XEP-0313: Relax 'communicating archive ID' requirement
The requirement keyword changed in this commit stems from the time when this XEP used <archived id='...' by='...' /> instead of XEP-0359.

The 'MUST NOT' requirement keyword implies that something will go horribly wrong, if the ID is sent. This is not the case here.

Sharing the stanza-id on a need-to-know basis is a sensible default, but should not be a requirement.