Fix shortname of XEP-0315: Data Forms XML Element
The shortname was 'media-element', which is already taken by XEP-0221:
Data Forms Media Element. It is likley that the shortname of XEP-0315
was errorneously set to 'media-element' caused by the author using
XEP-0221 as template for XEP-0315.

Shortnames of XEPs should be unique, hence this changes the shortname
of XEP-0315 from 'media-element' to 'xml-element'.