XEP-0393: make span rules match examples
Previously the styling directive rules indicated that the opening
styling directive must be preceded by whitespace or be the beginning of
the line, however, this meant that a later example was incorrect as it
showed bold and monospace text starting with *` (the second styling
directive is preceded by the first, not whitespace). Allowing styling
directives to immediately follow other styling directives fixes this
oversight.