XEP-0384: rephrase 'session termination' to 'opting-out'
Sessions are device-to-device relationships. When we talk about
'opting-out' we want to communicate a _preference_ for an entire account
not just on specific device-to-device relationship.

The actual double ratchet sessions will also remain intact. So it is not
reversing what ever we did in 'Starting a session' but just switching back
from omemo to plain. (When we currently do that manuelly we also wouldn’t
touch the actual underlying sessions)

I would like to note at this point that we don’t have a way of 'opting-in'
either. Nowhere do we say that a client MUST use OMEMO if they discover
that the other side has bundles. Therefor I find it strange that the opt-out
was a MUST and i weakend it to SHOULD. (we don’t say that a client SHOULD start
omemo when it discovers support either so I even find a SHOULD too strong)