XEP-0389: major updates to add more flexibility
Zash reached out to me to ask if extensible IBR could be made compatible
with XEP-0401: Easy User Onboarding. While working on a solution to this
I decided it would also be nice if it could also allow the user to pick
between several flows (eg. "Verify by Phone" vs  "Verify by SMS").
After working this up I came to realize that the current version of the
spec could not be used to replace XEP-0077 as it is used by XEP-0045 and
other XEPs because I wasn't using routeable stanzas that could be sent
to components and had no way to discover support other than during
stream feature negotation.

This patch set turned into an overhaul of the proposal to fix all these
problems and move the spec back out of experimental.