XEP-0060: Disallow = and ; in NodeIDs to allow use in URIs and refer to PRECIS Stringprep
The reason is that currently it is possible to forge possibly dangerous PubSub node IDs.
Let's take a node with ID catpictures;node=dogpictures as an example.
The URI for subscribing to above node might be `xmpp:pubsub.example.org?pubsub;action=subscribe;node=catpictures;node=dogpictures` which a client might interpret as `xmpp:pubsub.example.org?pubsub;action=subscribe;node=dogpictures`, which would subscribe the client to a different node.

Additionally Florian Schmaus noted that the sentence 'the pubsub service MUST ensure that the NodeID conforms to the Resourceprep profile of Stringprep as described in RFC 3920' is wrong and that PRECIS is used instead.
RFC 7622 §3.4 states that
'The resourcepart of a JID is an instance of the OpaqueString profile of the PRECIS FreeformClass, which is specified in [RFC7613]. The rules and considerations provided in that specification MUST be applied to XMPP resourceparts.'
Thanks for the input!