XEP-0004: Clarify that <reported/> must appear first
It is important for form consuming entities which validate the form
field types that the <reported/> element appears before any <item/>
element in result forms. This is because the <reported/> element
contains the information about the field type of the fields. So if a
form has any <item/> before <reported/>, valididating entites would
need two passes to parse the form: the first pass to parse <reported/>
and the second one to parse the fields in <item/>s.

I consider this a non-breaking change, and just a clarification of the
XEP authors intention.