Update xep-iwe.xml
I really like this proposal and have been wanting something similar for a while.

# `/ws` -> `/.well-known/xmpp-websocket`

In deployments, `/ws` is too generic and the domain might already have something running there for :443.

# `5443` -> `5280`

Quoting Marvin W

> - ws://...:5443/ws makes no sense. Port 5443 is obviously a reference to
HTTPS 443 port which is TLS encrypted, so you shouldn't make non-TLS
connections to 5443. If any, use something like port 5080 or 5280.

# server vendors

I'm not happy with the wording, suggestions?