xep.xsl: produce BibLaTeX entry
I am tired of seeing all those badly or misreferenced XEPs out
there. This produces a high-quality BibLaTeX entry that follows the
recommendations of the biblatex Package and
draft-carpenter-rfc-citation-recs-01 § 5.2 [1].

It also uses proper 'date' attribution, from the first date in the XEP
history to the date of the latest revision entry.

1: https://tools.ietf.org/html/draft-carpenter-rfc-citation-recs-01#section-5.2