XEP-0280: Remove the removal of the `<private/>` element
The requirement to strip `<private/>` by the sending server was in the
XEP from day 1. It was later changed from "MUST" to "SHOULD" and from
"sending server" to "receiving server", so that the latter could also
prevent normal message routing.

As this behavior only ever affected the `<private/>` element and not the
`<no-copy/>` hint (introduced in 2017), which should be treated in a
similar but not equal way, and there is a security benefit in letting
the receiving client know that message routing was manipulated, it makes
sense to remove this requirement.

As there is no negative effect of keeping another element in delivered
messages, nobody complained about `<no-copy/>` not being stripped, and
bumping Carbons today would be rather unfortunate, this is done without
a namespace bump, despite "breaking" the specification.