XEP-0363: Incorporate Council feedback
The wording about allowed headers was deemed unclear, so it was
clarified that only allowed headers should be forwarded.

In addition, there were concerns about implementation support for
preserving the relative order of different headers. Considering that
that's unlikely to be a use-case, the wording was changed to not require
that, but only preserving the order of values for the same header.