Make XEP superseded more visible in HTML output
When a XEP has been superseded (or supersedes something else) it's not
always immediately obvious in the HTML output. It exists at the bottom
of the page in the appendices, but this requires a lot of digging, and
in the status message, but this is a big red blob of text that we don't
expect to differ from XEP to XEP and is harder to parse through than the
information we want being in a table.

This adds the information to a third place, the table in the header to
make it more obvious.