Linkify revision history to attic
Comparing versions is a thing you may want to do when dealing with a
more or less living standard. Locating these versions in the attic is
often tedious since there are so many XEPs and revisions.

Linking them from the XEP revision history makes this easier, at least
where the relevant revision does exist in the attic, not all do, and
some are called jep rather than xep for historical reasons.