Add license for bibliographic data (third attempt)
We were asked [1] to state that the XEP bibliographic data is openly
available and free of charge if we want it to be consumed by third
parties.

This weas previously merged as 747738804a2d ("Add license for
bibliographic data"), but unfortunately later reverted in
59b2a5ca15c9 ("Revert "Add license for bibliographic data"",
https://github.com/xsf/xeps/pull/1221). However, the fix for the
broken docker build is trivial: simply also copy refs-LICENSE into the
container. This commit does that.

Fixes #1219.

1: https://github.com/ietf-tools/bibxml-service/issues/302#issuecomment-1274686153