XEP-0317: New revision 0.2
Despite being a popular concept that the community keeps coming back to, this
XEP has lingered in an unfinished state for some time. This small update fixes
some of the fundamental issues.

The current version did not actually specify a syntax for hats, as this was
still under discussion at the time. Based on discussions with other community
members, and personal implementation experience, this update specifies the
syntax to be used.