tools/2xep.lua: Fix <code> elements having unsupported attributes
The XEP DTD only supports a single attribute for <code>, the 'caption'
attribute. I don't know if/how it's possible to attach a caption in markdown,
so I'm leaving that as unsupported. That means no attributes are supported,
and therefore I've removed them.

Usage of ```xml is common, so I've also added code to handle those and convert
them to example blocks.