Makefile: add 'check' target, consolidating xmllint invocations
Previously the html and pdf targets would also verify the input via
xmllint. This caused some code duplication and overloaded the targets
which an arguably unrelated task (verification).

Furthermore, something changed and the

! xmllint --nonet --noout --noent --loaddtd --xpath "//img/@src[not(starts-with(., 'data:'))]" $< 2>/dev/null && true

no fails. It seems that xmllint does no longer (?) report an error
exit status if the XPath result set is empty [1]. Therfore, this rule
is currently commented out, which fixed #1316.

1: https://gitlab.gnome.org/GNOME/libxml2/-/issues/673