XEP-0428: Change schema int types to xs:unsignedInt
The previous type was xs:integer, which allows for arbitrary large
integers, and also negative integers. Negative values do not make sense
when referencing characters and limiting the size of the integer to 32
bit is, I think, reasonable as that would still allow referencing in 4
GiB of text (and that should probably exceed any stanza size limit).