XEP-0054: Updated error cases to be compatible with RFC 6121
In section 3.3 of XEP-0054 it is defined that the same error condition MUST be used in both a case:

- where a VCard request is addressed to an entity that does not exist, and;
- where a VCard request is addressed at an entity for which no VCard is available.

RFC 6121 section 8.5.1 defines:
> If the user account identified by the 'to' attribute does not exist, how the stanza is processed depends on the stanza type. For an IQ stanza, the server MUST return a `<service-unavailable/>` stanza error to the sender.

Prior to this change, both `<service-unavailable/>` and `<item-not-found/>` were offered as suitable conditions to be used in the error scenario defined in section 3.3. For this XEP to be more compatible with RFC 6121, only <service-unavailable> can be used.

This commit drops the `<item-not-found>` condition from the scenario, and elevates usage of `<service-unavailable>` to a MUST. It adds a note on backwards compatibility.