XEP-0045: Introduce section on role/affilition change notifications
The mechanisms in which occupants are notified of a role or affiliation change is quite similar throughout the specification, but is not consistently documented.

This commit introduces a new section, describing the pre-existing MUST of using a presence stanza when to affected user is in the room, as well as the pre-existing MAY of usinga message stanza when the user is not.