XEP-0045: Improve 'Service Removes Non-Member' example
The pre-existing example appears to show two mutually exclusive acceptable stanzas. This can be confused for a requirement to send two stanzas. In this commit, the example is improved by removing one of the variants. A small textual change has been applied to define the optionality of the 'actor' element.