XEP-0198: Clarify server enabling SM without requested resumption
When a client requests stream management to be enabled and asks for stream resumption, it's not unreasonable for people to assume that the server may 'fail' the request if it doesn't offer the requested resumption. The pre-existing test does allow the server to enable SM _without_ the requested resumption feature, but isn't overly clear.

This commit adds an example clarifying this scenario.