XEP-0045: Replace inappropriate RFC 2119 key word usage in §9.7
The 'MAY' key word, per RFC 2119, is to be used to flag a 'truly optional' item. In this text, it seems to be used to restrict the item to certain characteristics intead. This is an improper use of the key word.

Further down the section, the scenario is further specified, using more appropriate key word usage.