XEP-0004: Empty and absent values
The pre-existing specification does not differentiate between empty or absent values. As the XEP has reached the 'final' status, such definitions can not be added to this XEP now.

This commit doesn't (intend to) add new definitions. It merely documents consequences of what is the current final spec.

The intention of this change is to make explicit that users of this XEP cannot expect this distinction to be present (unless it is explicitly defined in dependant XEPs). As a result implementations can be expected to use a variety of representations.