XEP-0060: Move Discovery of hierarchy in Nodes to XEP-0248
Moves section 5.2 "Discover Nodes" from XEP-0060 (which has text and examples that are based on hierarchies / collections) to XEP-0248, replacing therein the section of the same name and number (which did _not_ explictly use hierarchy in its example).

Section 5.2 in XEP-0060 is replaces with text and examples that do not use any hierarchy.

This change makes explicit a requirement (MUST) to enable entities to discover the nodes in the service by means of the Service Discovery protocol. Prior to this commit, this requirement arguably applied only to a hierarchy of nodes.