Move process for changing node of type leaf to collection from XEP-0060 to XEP-0248
Prior to this change XEP-0060 described changing a leaf to a collection, while XEP-0248 described changing a collection to a leaf (but uses generic language that is easily interpreted as if node type changes are not allowable at all). By having both processes be specified in XEP-0248, there is a lot less room for ambiguity (while it also further reduces complexity in XEP-0060).