XEP-0060 & XEP-0248: Default node-type specific configuration
This adds an explicit section that describes how default configuration for a particular node-type can be obtained.

The mechanism described herein was already defined in the schema for XEP-0060 (although in the wrong schema, which is corrected in this commit). Implementations are known to exist in the wild.

This commit follows a discussion in the standards MUC on september 10 and 11, 2025. The logs are available at https://logs.xmpp.org/xsf/2025-09-11#2025-09-11-8bd371fea14c9595