XEP-0060 / XEP-0248: Improve separation of 'collections'
* XEP-0060: Move Discovery of hierarchy in Nodes to XEP-0248

Moves section 5.2 "Discover Nodes" from XEP-0060 (which has text and examples that are based on hierarchies / collections) to XEP-0248, replacing therein the section of the same name and number (which did _not_ explictly use hierarchy in its example).

Section 5.2 in XEP-0060 is replaces with text and examples that do not use any hierarchy.

This change makes explicit a requirement (MUST) to enable entities to discover the nodes in the service by means of the Service Discovery protocol. Prior to this commit, this requirement arguably applied only to a hierarchy of nodes.

* XEP-0060: Recommend JID+NodeID addressing when 'node' attribute is available

Prevents conflicts in protocols that provide a 'node' attribute, where the absense of a 'node' value is significant (such as to distinguish between 'Node Discovery' and 'Item Discovery').

* Move the Service Discovery Identity type 'collection' from XEP-0060 to XEP-0248

* Move collection-specific Service Discovery Features from XEP-0060 to XEP-0248

* Move process for changing node of type leaf to collection from XEP-0060 to XEP-0248

Prior to this change XEP-0060 described changing a leaf to a collection, while XEP-0248 described changing a collection to a leaf (but uses generic language that is easily interpreted as if node type changes are not allowable at all). By having both processes be specified in XEP-0248, there is a lot less room for ambiguity (while it also further reduces complexity in XEP-0060).

* XEP-0060: Remove reference to non-existing process

XEP-0060 should no longer reference to XEP-0248 for a 'process to retrieve the default subscription configuration options for collection nodes'. Such process is not explicitly defined in XEP-0248 (meaning that the process is equal to that in XEP-0060), meaning that the reference is more likely to cause confusion than clarity. Although additional configuration options (but not a new process) are provided in XEP-0248, it is perfectly reasonable for readers to expect that type-specific configuration is to be found in type-specific XEPs.

* Move requirement to support pubsub#notify_config for collection nodes from XEP-0060 to XEP-0248

The hard requirement that is specific to collection nodes should be defined in the XEP for collection nodes.

* XEP-0060 Remove collection-specific pubsub#node_config FORM_TYPE field standardization

Removes from XEP-0060 field standardization that already exists in XEP-0248.

* XEP-0060: Remove pubsub#collection SHIM header registration as its registration is duplicated in XEP-0248

* XEP-0060: Apply Council feedback

- Fixed 'therefor' typo
- Changed disco#items support from MUST to SHOULD (with conditions)

---------

Co-authored-by: Daniel Gultsch <daniel@gultsch.de>