XEP-0060: Unmarking 'subscribe' and 'publish' features as mandatory
Sections 6.1.3.10 and 7.1.3.2 contain normative text that describe how 'subscribe' and 'publish' functionality can be unsupported by a service. From this, it follows that these are features that are not _required_ to be implemented by a service. In section 10 'Feature Summary', these features should by this logic not be defined as REQUIRED.